split cfgdlg part 2
