+++ trelby/cfgdlg.py
@@ -1,17 +1,21 @@
-import functools
 import os.path
 
 import wx
 
-import trelby.config as config
 import trelby.gutil as gutil
 import trelby.misc as misc
-import trelby.screenplay as screenplay
 import trelby.truetype as truetype
 import trelby.util as util
 
 from trelby.configpages.globalaboutpanel import GlobalAboutPanel
 from trelby.configpages.scriptaboutpanel import ScriptAboutPanel
+from trelby.configpages.elementspanel import ElementsPanel
+from trelby.configpages.elementsglobalpanel import ElementsGlobalPanel
+from trelby.configpages.colorspanel import ColorsPanel
+from trelby.configpages.paperpanel import PaperPanel
+from trelby.configpages.formattingpanel import FormattingPanel
+from trelby.configpages.stringspanel import StringsPanel
+from trelby.configpages.pdfpanel import PDFPanel
 
 # stupid hack to get correct window modality stacking for dialogs
 cfgFrame = None
@@ -328,615 +332,6 @@ class DisplayPanel(wx.Panel):
     def cfg2gui(self):
         self.spacingEntry.SetValue(self.cfg.fontYdelta)
         self.pbRb.SetSelection(self.cfg.pbi)
-
-
-class ElementsPanel(wx.Panel):
-    def __init__(self, parent, id, cfg):
-        wx.Panel.__init__(self, parent, id)
-        self.cfg = cfg
-
-        vsizer = wx.BoxSizer(wx.VERTICAL)
-
-        hsizer = wx.BoxSizer(wx.HORIZONTAL)
-
-        hsizer.Add(
-            wx.StaticText(self, -1, "Element:"),
-            0,
-            wx.ALIGN_CENTER_VERTICAL | wx.RIGHT,
-            10,
-        )
-
-        self.elementsCombo = wx.ComboBox(self, -1, style=wx.CB_READONLY)
-
-        for t in config.getTIs():
-            self.elementsCombo.Append(t.name, t.lt)
-
-        hsizer.Add(self.elementsCombo, 0)
-
-        vsizer.Add(hsizer, 0, wx.EXPAND)
-
-        vsizer.Add(wx.StaticLine(self, -1), 0, wx.EXPAND | wx.TOP | wx.BOTTOM, 10)
-
-        hsizer = wx.BoxSizer(wx.HORIZONTAL)
-
-        hsizer.Add(self.addTextStyles("Screen", "screen", self))
-        hsizer.Add(self.addTextStyles("Print", "export", self), 0, wx.LEFT, 10)
-
-        vsizer.Add(hsizer, 0, wx.BOTTOM, 10)
-
-        gsizer = wx.FlexGridSizer(2, 2, 5, 0)
-
-        gsizer.Add(
-            wx.StaticText(self, -1, "Empty lines / 10 before:"),
-            0,
-            wx.ALIGN_CENTER_VERTICAL | wx.RIGHT,
-            10,
-        )
-
-        tmp = wx.SpinCtrl(self, -1)
-        tmp.SetRange(
-            *self.cfg.getType(screenplay.ACTION).cvars.getMinMax("beforeSpacing")
-        )
-        self.Bind(wx.EVT_SPINCTRL, self.OnMisc, id=tmp.GetId())
-        tmp.Bind(wx.EVT_KILL_FOCUS, self.OnKillFocus)
-        gsizer.Add(tmp)
-        self.beforeSpacingEntry = tmp
-
-        gsizer.Add(
-            wx.StaticText(self, -1, "Empty lines / 10 between:"),
-            0,
-            wx.ALIGN_CENTER_VERTICAL | wx.RIGHT,
-            10,
-        )
-
-        tmp = wx.SpinCtrl(self, -1)
-        tmp.SetRange(
-            *self.cfg.getType(screenplay.ACTION).cvars.getMinMax("intraSpacing")
-        )
-        self.Bind(wx.EVT_SPINCTRL, self.OnMisc, id=tmp.GetId())
-        tmp.Bind(wx.EVT_KILL_FOCUS, self.OnKillFocus)
-        gsizer.Add(tmp)
-        self.intraSpacingEntry = tmp
-
-        vsizer.Add(gsizer, 0, wx.BOTTOM, 20)
-
-        gsizer = wx.FlexGridSizer(2, 3, 5, 0)
-
-        gsizer.Add(
-            wx.StaticText(self, -1, "Indent:"),
-            0,
-            wx.ALIGN_CENTER_VERTICAL | wx.RIGHT,
-            10,
-        )
-
-        self.indentEntry = wx.SpinCtrl(self, -1)
-        self.indentEntry.SetRange(
-            *self.cfg.getType(screenplay.ACTION).cvars.getMinMax("indent")
-        )
-        self.Bind(wx.EVT_SPINCTRL, self.OnMisc, id=self.indentEntry.GetId())
-        self.indentEntry.Bind(wx.EVT_KILL_FOCUS, self.OnKillFocus)
-        gsizer.Add(self.indentEntry, 0)
-
-        gsizer.Add(
-            wx.StaticText(self, -1, "characters (10 characters" " = 1 inch)"),
-            0,
-            wx.ALIGN_CENTER_VERTICAL | wx.LEFT,
-            10,
-        )
-
-        gsizer.Add(
-            wx.StaticText(self, -1, "Width:"),
-            0,
-            wx.ALIGN_CENTER_VERTICAL | wx.RIGHT,
-            10,
-        )
-
-        self.widthEntry = wx.SpinCtrl(self, -1)
-        self.widthEntry.SetRange(
-            *self.cfg.getType(screenplay.ACTION).cvars.getMinMax("width")
-        )
-        self.Bind(wx.EVT_SPINCTRL, self.OnMisc, id=self.widthEntry.GetId())
-        self.widthEntry.Bind(wx.EVT_KILL_FOCUS, self.OnKillFocus)
-        gsizer.Add(self.widthEntry, 0)
-
-        gsizer.Add(
-            wx.StaticText(self, -1, "characters (10 characters" " = 1 inch)"),
-            0,
-            wx.ALIGN_CENTER_VERTICAL | wx.LEFT,
-            10,
-        )
-
-        vsizer.Add(gsizer, 0, wx.BOTTOM, 20)
-
-        util.finishWindow(self, vsizer, center=False)
-
-        self.Bind(wx.EVT_COMBOBOX, self.OnElementCombo, id=self.elementsCombo.GetId())
-
-        self.elementsCombo.SetSelection(0)
-        self.OnElementCombo()
-
-    def addTextStyles(self, name, prefix, parent):
-        hsizer = wx.StaticBoxSizer(wx.StaticBox(parent, -1, name), wx.HORIZONTAL)
-
-        gsizer = wx.FlexGridSizer(2, 2, 0, 10)
-
-        # wxGTK adds way more space by default than wxMSW between the
-        # items, have to adjust for that
-        pad = 0
-        if misc.isWindows:
-            pad = 5
-
-        self.addCheckBox("Caps", prefix, parent, gsizer, pad)
-        self.addCheckBox("Italic", prefix, parent, gsizer, pad)
-        self.addCheckBox("Bold", prefix, parent, gsizer, pad)
-        self.addCheckBox("Underlined", prefix, parent, gsizer, pad)
-
-        hsizer.Add(gsizer, 0, wx.EXPAND)
-
-        return hsizer
-
-    def addCheckBox(self, name, prefix, parent, sizer, pad):
-        cb = wx.CheckBox(parent, -1, name)
-        self.Bind(wx.EVT_CHECKBOX, self.OnStyleCb, id=cb.GetId())
-        sizer.Add(cb, 0, wx.TOP, pad)
-        setattr(self, prefix + name + "Cb", cb)
-
-    def OnKillFocus(self, event):
-        self.OnMisc()
-
-        # if we don't call this, the spin entry on wxGTK gets stuck in
-        # some weird state
-        event.Skip()
-
-    def OnElementCombo(self, event=None):
-        self.lt = self.elementsCombo.GetClientData(self.elementsCombo.GetSelection())
-        self.cfg2gui()
-
-    def OnStyleCb(self, event):
-        tcfg = self.cfg.types[self.lt]
-
-        tcfg.screen.isCaps = self.screenCapsCb.GetValue()
-        tcfg.screen.isItalic = self.screenItalicCb.GetValue()
-        tcfg.screen.isBold = self.screenBoldCb.GetValue()
-        tcfg.screen.isUnderlined = self.screenUnderlinedCb.GetValue()
-
-        tcfg.export.isCaps = self.exportCapsCb.GetValue()
-        tcfg.export.isItalic = self.exportItalicCb.GetValue()
-        tcfg.export.isBold = self.exportBoldCb.GetValue()
-        tcfg.export.isUnderlined = self.exportUnderlinedCb.GetValue()
-
-    def OnMisc(self, event=None):
-        tcfg = self.cfg.types[self.lt]
-
-        tcfg.beforeSpacing = util.getSpinValue(self.beforeSpacingEntry)
-        tcfg.intraSpacing = util.getSpinValue(self.intraSpacingEntry)
-        tcfg.indent = util.getSpinValue(self.indentEntry)
-        tcfg.width = util.getSpinValue(self.widthEntry)
-
-    def cfg2gui(self):
-        tcfg = self.cfg.types[self.lt]
-
-        self.screenCapsCb.SetValue(tcfg.screen.isCaps)
-        self.screenItalicCb.SetValue(tcfg.screen.isItalic)
-        self.screenBoldCb.SetValue(tcfg.screen.isBold)
-        self.screenUnderlinedCb.SetValue(tcfg.screen.isUnderlined)
-
-        self.exportCapsCb.SetValue(tcfg.export.isCaps)
-        self.exportItalicCb.SetValue(tcfg.export.isItalic)
-        self.exportBoldCb.SetValue(tcfg.export.isBold)
-        self.exportUnderlinedCb.SetValue(tcfg.export.isUnderlined)
-
-        # stupid wxwindows/wxpython displays empty box if the initial
-        # value is zero if we don't do this...
-        self.beforeSpacingEntry.SetValue(5)
-        self.intraSpacingEntry.SetValue(5)
-        self.indentEntry.SetValue(5)
-
-        self.beforeSpacingEntry.SetValue(tcfg.beforeSpacing)
-        self.intraSpacingEntry.SetValue(tcfg.intraSpacing)
-        self.indentEntry.SetValue(tcfg.indent)
-        self.widthEntry.SetValue(tcfg.width)
-
-
-class ColorsPanel(wx.Panel):
-    def __init__(self, parent, id, cfg):
-        wx.Panel.__init__(self, parent, id)
-        self.cfg = cfg
-
-        vsizer = wx.BoxSizer(wx.VERTICAL)
-
-        hsizer = wx.BoxSizer(wx.HORIZONTAL)
-
-        self.colorsLb = wx.ListBox(self, -1, size=(300, 250))
-
-        tmp = list(self.cfg.cvars.color.values())
-        tmp = sorted(
-            tmp, key=functools.cmp_to_key(lambda c1, c2: cmpfunc(c1.descr, c2.descr))
-        )
-
-        for it in tmp:
-            self.colorsLb.Append(it.descr, it.name)
-
-        hsizer.Add(self.colorsLb, 1)
-
-        vsizer.Add(hsizer, 0, wx.EXPAND | wx.BOTTOM, 10)
-
-        hsizer = wx.BoxSizer(wx.HORIZONTAL)
-
-        vsizer2 = wx.BoxSizer(wx.VERTICAL)
-
-        btn = wx.Button(self, -1, "Change")
-        self.Bind(wx.EVT_BUTTON, self.OnChangeColor, id=btn.GetId())
-        vsizer2.Add(btn, 0, wx.BOTTOM, 10)
-
-        btn = wx.Button(self, -1, "Restore default")
-        self.Bind(wx.EVT_BUTTON, self.OnDefaultColor, id=btn.GetId())
-        vsizer2.Add(btn)
-
-        hsizer.Add(vsizer2, 0, wx.ALIGN_CENTER_VERTICAL | wx.RIGHT, 10)
-
-        self.colorSample = misc.MyColorSample(self, -1, size=wx.Size(200, 50))
-        hsizer.Add(self.colorSample, 1, wx.EXPAND)
-
-        vsizer.Add(hsizer, 0, wx.EXPAND | wx.BOTTOM, 10)
-
-        self.useCustomElemColors = wx.CheckBox(self, -1, "Use per-element-type colors")
-        self.Bind(wx.EVT_CHECKBOX, self.OnMisc, id=self.useCustomElemColors.GetId())
-        vsizer.Add(self.useCustomElemColors)
-
-        util.finishWindow(self, vsizer, center=False)
-
-        self.Bind(wx.EVT_LISTBOX, self.OnColorLb, id=self.colorsLb.GetId())
-        self.colorsLb.SetSelection(0)
-        self.OnColorLb()
-
-    def OnColorLb(self, event=None):
-        self.color = self.colorsLb.GetClientData(self.colorsLb.GetSelection())
-        self.cfg2gui()
-
-    def OnChangeColor(self, event):
-        cd = wx.ColourData()
-        cd.SetColour(getattr(self.cfg, self.color).toWx())
-        dlg = wx.ColourDialog(self, cd)
-        dlg.SetTitle(self.colorsLb.GetStringSelection())
-        if dlg.ShowModal() == wx.ID_OK:
-            setattr(
-                self.cfg,
-                self.color,
-                util.MyColor.fromWx(dlg.GetColourData().GetColour()),
-            )
-        dlg.Destroy()
-
-        self.cfg2gui()
-
-    def OnDefaultColor(self, event):
-        setattr(self.cfg, self.color, self.cfg.cvars.getDefault(self.color))
-        self.cfg2gui()
-
-    def OnMisc(self, event=None):
-        self.cfg.useCustomElemColors = self.useCustomElemColors.GetValue()
-
-    def cfg2gui(self):
-        self.useCustomElemColors.SetValue(self.cfg.useCustomElemColors)
-
-        self.colorSample.SetBackgroundColour(getattr(self.cfg, self.color).toWx())
-        self.colorSample.Refresh()
-
-
-class PaperPanel(wx.Panel):
-    def __init__(self, parent, id, cfg):
-        wx.Panel.__init__(self, parent, id)
-        self.cfg = cfg
-
-        self.blockEvents = 1
-
-        self.paperSizes = {
-            "A4": (210.0, 297.0),
-            "Letter": (215.9, 279.4),
-            "Custom": (1.0, 1.0),
-        }
-
-        vsizer = wx.BoxSizer(wx.VERTICAL)
-
-        gsizer = wx.FlexGridSizer(3, 2, 5, 5)
-
-        gsizer.Add(
-            wx.StaticText(self, -1, "Type:"), 0, wx.ALIGN_CENTER_VERTICAL | wx.RIGHT, 10
-        )
-
-        self.paperCombo = wx.ComboBox(self, -1, style=wx.CB_READONLY)
-
-        for k, v in list(self.paperSizes.items()):
-            self.paperCombo.Append(k, v)
-
-        gsizer.Add(self.paperCombo)
-
-        gsizer.Add(wx.StaticText(self, -1, "Width:"), 0, wx.ALIGN_CENTER_VERTICAL)
-        hsizer = wx.BoxSizer(wx.HORIZONTAL)
-        self.widthEntry = wx.TextCtrl(self, -1)
-        hsizer.Add(self.widthEntry)
-        hsizer.Add(
-            wx.StaticText(self, -1, "mm"), 0, wx.ALIGN_CENTER_VERTICAL | wx.LEFT, 5
-        )
-        gsizer.Add(hsizer)
-
-        gsizer.Add(wx.StaticText(self, -1, "Height:"), 0, wx.ALIGN_CENTER_VERTICAL)
-        hsizer = wx.BoxSizer(wx.HORIZONTAL)
-        self.heightEntry = wx.TextCtrl(self, -1)
-        hsizer.Add(self.heightEntry)
-        hsizer.Add(
-            wx.StaticText(self, -1, "mm"), 0, wx.ALIGN_CENTER_VERTICAL | wx.LEFT, 5
-        )
-        gsizer.Add(hsizer)
-
-        vsizer.Add(gsizer, 0, wx.BOTTOM, 10)
-
-        bsizer = wx.StaticBoxSizer(wx.StaticBox(self, -1, "Margins"), wx.HORIZONTAL)
-
-        gsizer = wx.FlexGridSizer(4, 5, 5, 5)
-
-        self.addMarginCtrl("Top", self, gsizer)
-        self.addMarginCtrl("Bottom", self, gsizer)
-        self.addMarginCtrl("Left", self, gsizer)
-        self.addMarginCtrl("Right", self, gsizer)
-
-        bsizer.Add(gsizer, 0, wx.EXPAND | wx.ALL, 10)
-
-        vsizer.Add(bsizer, 0, wx.BOTTOM, 10)
-
-        vsizer.Add(wx.StaticText(self, -1, "(1 inch = 25.4 mm)"), 0, wx.LEFT, 25)
-
-        self.linesLabel = wx.StaticText(self, -1, "")
-
-        # wxwindows doesn't recalculate sizer size correctly at startup so
-        # set initial text
-        self.setLines()
-
-        vsizer.Add(self.linesLabel, 0, wx.TOP, 20)
-
-        util.finishWindow(self, vsizer, center=False)
-
-        ptype = "Custom"
-        for k, v in list(self.paperSizes.items()):
-            if self.eqFloat(self.cfg.paperWidth, v[0]) and self.eqFloat(
-                self.cfg.paperHeight, v[1]
-            ):
-                ptype = k
-
-        idx = self.paperCombo.FindString(ptype)
-        if idx != -1:
-            self.paperCombo.SetSelection(idx)
-
-        self.Bind(wx.EVT_COMBOBOX, self.OnPaperCombo, id=self.paperCombo.GetId())
-        self.OnPaperCombo(None)
-
-        self.Bind(wx.EVT_TEXT, self.OnMisc, id=self.widthEntry.GetId())
-        self.Bind(wx.EVT_TEXT, self.OnMisc, id=self.heightEntry.GetId())
-
-        self.cfg2mm()
-        self.cfg2inch()
-
-        self.blockEvents -= 1
-
-    def eqFloat(self, f1, f2):
-        return round(f1, 2) == round(f2, 2)
-
-    def addMarginCtrl(self, name, parent, sizer):
-        sizer.Add(wx.StaticText(parent, -1, name + ":"), 0, wx.ALIGN_CENTER_VERTICAL)
-
-        entry = wx.TextCtrl(parent, -1)
-        sizer.Add(entry, 0)
-        label = wx.StaticText(parent, -1, "mm")
-        sizer.Add(label, 0, wx.ALIGN_CENTER_VERTICAL)
-
-        entry2 = wx.TextCtrl(parent, -1)
-        sizer.Add(entry2, 0, wx.LEFT, 20)
-        label2 = wx.StaticText(parent, -1, "inch")
-        sizer.Add(label2, 0, wx.ALIGN_CENTER_VERTICAL)
-
-        setattr(self, name.lower() + "EntryMm", entry)
-        setattr(self, name.lower() + "EntryInch", entry2)
-
-        self.Bind(wx.EVT_TEXT, self.OnMarginMm, id=entry.GetId())
-        self.Bind(wx.EVT_TEXT, self.OnMarginInch, id=entry2.GetId())
-
-    def doForcedUpdate(self):
-        self.setLines()
-
-    def setLines(self):
-        self.cfg.recalc(False)
-        self.linesLabel.SetLabel("Lines per page: %d" % self.cfg.linesOnPage)
-
-    def OnPaperCombo(self, event):
-        w, h = self.paperCombo.GetClientData(self.paperCombo.GetSelection())
-
-        ptype = self.paperCombo.GetStringSelection()
-
-        if ptype == "Custom":
-            self.widthEntry.Enable(True)
-            self.heightEntry.Enable(True)
-            w = self.cfg.paperWidth
-            h = self.cfg.paperHeight
-        else:
-            self.widthEntry.Disable()
-            self.heightEntry.Disable()
-
-        self.widthEntry.SetValue(str(w))
-        self.heightEntry.SetValue(str(h))
-
-        self.setLines()
-
-    def OnMisc(self, event):
-        if self.blockEvents > 0:
-            return
-
-        self.entry2float(self.widthEntry, "paperWidth")
-        self.entry2float(self.heightEntry, "paperHeight")
-
-        self.setLines()
-
-    def OnMarginMm(self, event):
-        if self.blockEvents > 0:
-            return
-
-        self.blockEvents += 1
-
-        self.entry2float(self.topEntryMm, "marginTop")
-        self.entry2float(self.bottomEntryMm, "marginBottom")
-        self.entry2float(self.leftEntryMm, "marginLeft")
-        self.entry2float(self.rightEntryMm, "marginRight")
-
-        self.setLines()
-
-        self.cfg2inch()
-
-        self.blockEvents -= 1
-
-    def OnMarginInch(self, event):
-        if self.blockEvents > 0:
-            return
-
-        self.blockEvents += 1
-
-        self.entry2float(self.topEntryInch, "marginTop", 25.4)
-        self.entry2float(self.bottomEntryInch, "marginBottom", 25.4)
-        self.entry2float(self.leftEntryInch, "marginLeft", 25.4)
-        self.entry2float(self.rightEntryInch, "marginRight", 25.4)
-
-        self.setLines()
-
-        self.cfg2mm()
-
-        self.blockEvents -= 1
-
-    def cfg2mm(self):
-        self.topEntryMm.SetValue(str(self.cfg.marginTop))
-        self.bottomEntryMm.SetValue(str(self.cfg.marginBottom))
-        self.leftEntryMm.SetValue(str(self.cfg.marginLeft))
-        self.rightEntryMm.SetValue(str(self.cfg.marginRight))
-
-    def cfg2inch(self):
-        self.topEntryInch.SetValue(str(self.cfg.marginTop / 25.4))
-        self.bottomEntryInch.SetValue(str(self.cfg.marginBottom / 25.4))
-        self.leftEntryInch.SetValue(str(self.cfg.marginLeft / 25.4))
-        self.rightEntryInch.SetValue(str(self.cfg.marginRight / 25.4))
-
-    def entry2float(self, entry, name, factor=1.0):
-        val = util.str2float(entry.GetValue(), 0.0) * factor
-        setattr(self.cfg, name, val)
-
-
-class FormattingPanel(wx.Panel):
-    def __init__(self, parent, id, cfg):
-        wx.Panel.__init__(self, parent, id)
-        self.cfg = cfg
-
-        vsizer = wx.BoxSizer(wx.VERTICAL)
-
-        vsizer.Add(
-            wx.StaticText(
-                self,
-                -1,
-                "Leave at least this many lines at the end of a page when\n"
-                "breaking in the middle of an element:",
-            ),
-            0,
-            wx.BOTTOM,
-            5,
-        )
-
-        gsizer = wx.FlexGridSizer(2, 2, 5, 0)
-
-        self.addSpin("action", "Action:", self, gsizer, "pbActionLines")
-        self.addSpin("dialogue", "Dialogue", self, gsizer, "pbDialogueLines")
-
-        vsizer.Add(gsizer, 0, wx.LEFT, 10)
-
-        hsizer = wx.BoxSizer(wx.HORIZONTAL)
-        self.addSpin("fontSize", "Font size:", self, hsizer, "fontSize")
-        vsizer.Add(hsizer, 0, wx.TOP, 20)
-
-        vsizer.Add(wx.StaticText(self, -1, "Scene CONTINUEDs:"), 0, wx.TOP, 20)
-
-        hsizer = wx.BoxSizer(wx.HORIZONTAL)
-        self.sceneContinuedsCb = wx.CheckBox(self, -1, "Include,")
-        self.Bind(wx.EVT_CHECKBOX, self.OnMisc, id=self.sceneContinuedsCb.GetId())
-        hsizer.Add(self.sceneContinuedsCb, 0, wx.ALIGN_CENTER_VERTICAL | wx.LEFT, 10)
-
-        self.addSpin(
-            "sceneContinuedIndent", "indent:", self, hsizer, "sceneContinuedIndent"
-        )
-        hsizer.Add(
-            wx.StaticText(self, -1, "characters"),
-            0,
-            wx.ALIGN_CENTER_VERTICAL | wx.LEFT,
-            10,
-        )
-        vsizer.Add(hsizer, 0, wx.LEFT, 5)
-
-        self.scenesCb = wx.CheckBox(self, -1, "Include scene numbers")
-        self.Bind(wx.EVT_CHECKBOX, self.OnMisc, id=self.scenesCb.GetId())
-        vsizer.Add(self.scenesCb, 0, wx.TOP, 10)
-
-        # wxGTK adds way more space by default than wxMSW between the
-        # items, have to adjust for that
-        pad = 0
-        if misc.isWindows:
-            pad = 10
-
-        self.lineNumbersCb = wx.CheckBox(self, -1, "Show line numbers (debug)")
-        self.Bind(wx.EVT_CHECKBOX, self.OnMisc, id=self.lineNumbersCb.GetId())
-        vsizer.Add(self.lineNumbersCb, 0, wx.TOP, pad)
-
-        self.cfg2gui()
-
-        util.finishWindow(self, vsizer, center=False)
-
-    def addSpin(self, name, descr, parent, sizer, cfgName):
-        sizer.Add(
-            wx.StaticText(parent, -1, descr), 0, wx.ALIGN_CENTER_VERTICAL | wx.RIGHT, 10
-        )
-
-        entry = wx.SpinCtrl(parent, -1)
-        entry.SetRange(*self.cfg.cvars.getMinMax(cfgName))
-        self.Bind(wx.EVT_SPINCTRL, self.OnMisc, id=entry.GetId())
-        entry.Bind(wx.EVT_KILL_FOCUS, self.OnKillFocus)
-        sizer.Add(entry, 0)
-
-        setattr(self, name + "Entry", entry)
-
-    def OnKillFocus(self, event):
-        self.OnMisc()
-
-        # if we don't call this, the spin entry on wxGTK gets stuck in
-        # some weird state
-        event.Skip()
-
-    def OnMisc(self, event=None):
-        self.cfg.pbActionLines = util.getSpinValue(self.actionEntry)
-        self.cfg.pbDialogueLines = util.getSpinValue(self.dialogueEntry)
-        self.cfg.sceneContinueds = self.sceneContinuedsCb.GetValue()
-        self.cfg.sceneContinuedIndent = util.getSpinValue(
-            self.sceneContinuedIndentEntry
-        )
-        self.cfg.fontSize = util.getSpinValue(self.fontSizeEntry)
-        self.cfg.pdfShowSceneNumbers = self.scenesCb.GetValue()
-        self.cfg.pdfShowLineNumbers = self.lineNumbersCb.GetValue()
-
-    def cfg2gui(self):
-        # stupid wxwindows/wxpython displays empty box if the initial
-        # value is zero if we don't do this...
-        self.actionEntry.SetValue(5)
-        self.dialogueEntry.SetValue(5)
-        self.sceneContinuedIndentEntry.SetValue(5)
-
-        self.actionEntry.SetValue(self.cfg.pbActionLines)
-        self.dialogueEntry.SetValue(self.cfg.pbDialogueLines)
-        self.sceneContinuedsCb.SetValue(self.cfg.sceneContinueds)
-        self.sceneContinuedIndentEntry.SetValue(self.cfg.sceneContinuedIndent)
-        self.fontSizeEntry.SetValue(self.cfg.fontSize)
-        self.scenesCb.SetValue(self.cfg.pdfShowSceneNumbers)
-        self.lineNumbersCb.SetValue(self.cfg.pdfShowLineNumbers)
 
 
 class KeyboardPanel(wx.Panel):
@@ -1282,203 +677,6 @@ class MiscPanel(wx.Panel):
         self.splashTimeEntry.SetValue(self.cfg.splashTime)
 
 
-class ElementsGlobalPanel(wx.Panel):
-    def __init__(self, parent, id, cfg):
-        wx.Panel.__init__(self, parent, id)
-        self.cfg = cfg
-
-        vsizer = wx.BoxSizer(wx.VERTICAL)
-
-        hsizer = wx.BoxSizer(wx.HORIZONTAL)
-
-        hsizer.Add(
-            wx.StaticText(self, -1, "Element:"),
-            0,
-            wx.ALIGN_CENTER_VERTICAL | wx.RIGHT,
-            10,
-        )
-
-        self.elementsCombo = wx.ComboBox(self, -1, style=wx.CB_READONLY)
-
-        for t in config.getTIs():
-            self.elementsCombo.Append(t.name, t.lt)
-
-        hsizer.Add(self.elementsCombo, 0)
-
-        vsizer.Add(hsizer, 0, wx.EXPAND)
-
-        vsizer.Add(wx.StaticLine(self, -1), 0, wx.EXPAND | wx.TOP | wx.BOTTOM, 10)
-
-        gsizer = wx.FlexGridSizer(0, 2, 5, 0)
-
-        self.addTypeCombo("newEnter", "Enter creates", self, gsizer)
-        self.addTypeCombo("newTab", "Tab creates", self, gsizer)
-        self.addTypeCombo("nextTab", "Tab switches to", self, gsizer)
-        self.addTypeCombo("prevTab", "Shift+Tab switches to", self, gsizer)
-
-        vsizer.Add(gsizer)
-
-        util.finishWindow(self, vsizer, center=False)
-
-        self.Bind(wx.EVT_COMBOBOX, self.OnElementCombo, id=self.elementsCombo.GetId())
-
-        self.elementsCombo.SetSelection(0)
-        self.OnElementCombo()
-
-    def addTypeCombo(self, name, descr, parent, sizer):
-        sizer.Add(
-            wx.StaticText(parent, -1, descr + ":"),
-            0,
-            wx.ALIGN_CENTER_VERTICAL | wx.RIGHT,
-            10,
-        )
-
-        combo = wx.ComboBox(parent, -1, style=wx.CB_READONLY)
-
-        for t in config.getTIs():
-            combo.Append(t.name, t.lt)
-
-        sizer.Add(combo)
-
-        self.Bind(wx.EVT_COMBOBOX, self.OnMisc, id=combo.GetId())
-
-        setattr(self, name + "Combo", combo)
-
-    def OnElementCombo(self, event=None):
-        self.lt = self.elementsCombo.GetClientData(self.elementsCombo.GetSelection())
-        self.cfg2gui()
-
-    def OnMisc(self, event=None):
-        tcfg = self.cfg.types[self.lt]
-
-        tcfg.newTypeEnter = self.newEnterCombo.GetClientData(
-            self.newEnterCombo.GetSelection()
-        )
-        tcfg.newTypeTab = self.newTabCombo.GetClientData(
-            self.newTabCombo.GetSelection()
-        )
-        tcfg.nextTypeTab = self.nextTabCombo.GetClientData(
-            self.nextTabCombo.GetSelection()
-        )
-        tcfg.prevTypeTab = self.prevTabCombo.GetClientData(
-            self.prevTabCombo.GetSelection()
-        )
-
-    def cfg2gui(self):
-        tcfg = self.cfg.types[self.lt]
-
-        util.reverseComboSelect(self.newEnterCombo, tcfg.newTypeEnter)
-        util.reverseComboSelect(self.newTabCombo, tcfg.newTypeTab)
-        util.reverseComboSelect(self.nextTabCombo, tcfg.nextTypeTab)
-        util.reverseComboSelect(self.prevTabCombo, tcfg.prevTypeTab)
-
-
-class StringsPanel(wx.Panel):
-    def __init__(self, parent, id, cfg):
-        wx.Panel.__init__(self, parent, id)
-        self.cfg = cfg
-
-        # list of names. each name is both the name of a wx.TextCtrl in
-        # this class and the name of a string configuration variable in
-        # cfg.
-        self.items = []
-
-        vsizer = wx.BoxSizer(wx.VERTICAL)
-
-        gsizer = wx.FlexGridSizer(4, 2, 5, 0)
-
-        self.addEntry("strContinuedPageEnd", "(CONTINUED)", self, gsizer)
-        self.addEntry("strContinuedPageStart", "CONTINUED:", self, gsizer)
-        self.addEntry("strMore", "(MORE)", self, gsizer)
-        self.addEntry("strDialogueContinued", " (cont'd)", self, gsizer)
-
-        gsizer.AddGrowableCol(1)
-        vsizer.Add(gsizer, 0, wx.EXPAND)
-
-        self.cfg2gui()
-
-        util.finishWindow(self, vsizer, center=False)
-
-        for it in self.items:
-            self.Bind(wx.EVT_TEXT, self.OnMisc, id=getattr(self, it).GetId())
-
-    def addEntry(self, name, descr, parent, sizer):
-        sizer.Add(
-            wx.StaticText(parent, -1, descr), 0, wx.ALIGN_CENTER_VERTICAL | wx.RIGHT, 10
-        )
-
-        tmp = wx.TextCtrl(parent, -1)
-        sizer.Add(tmp, 0, wx.EXPAND | wx.ALIGN_CENTER_VERTICAL)
-
-        setattr(self, name, tmp)
-        self.items.append(name)
-
-    def OnMisc(self, event=None):
-        for it in self.items:
-            setattr(self.cfg, it, misc.fromGUI(getattr(self, it).GetValue()))
-
-    def cfg2gui(self):
-        for it in self.items:
-            getattr(self, it).SetValue(getattr(self.cfg, it))
-
-
-class PDFPanel(wx.Panel):
-    def __init__(self, parent, id, cfg):
-        wx.Panel.__init__(self, parent, id)
-        self.cfg = cfg
-
-        vsizer = wx.BoxSizer(wx.VERTICAL)
-
-        # wxGTK adds way more space by default than wxMSW between the
-        # items, have to adjust for that
-        pad = 0
-        if misc.isWindows:
-            pad = 10
-
-        self.includeTOCCb = self.addCb("Add table of contents", vsizer, pad)
-
-        self.showTOCCb = self.addCb("Show table of contents on PDF open", vsizer, pad)
-
-        self.openOnCurrentPageCb = self.addCb("Open PDF on current page", vsizer, pad)
-
-        self.removeNotesCb = self.addCb("Omit Note elements", vsizer, pad)
-
-        self.outlineNotesCb = self.addCb(
-            "  Draw rectangles around Note elements", vsizer, pad
-        )
-
-        self.marginsCb = self.addCb("Show margins (debug)", vsizer, pad)
-
-        self.cfg2gui()
-
-        util.finishWindow(self, vsizer, center=False)
-
-    def addCb(self, descr, sizer, pad):
-        ctrl = wx.CheckBox(self, -1, descr)
-        self.Bind(wx.EVT_CHECKBOX, self.OnMisc, id=ctrl.GetId())
-        sizer.Add(ctrl, 0, wx.TOP, pad)
-
-        return ctrl
-
-    def OnMisc(self, event=None):
-        self.cfg.pdfIncludeTOC = self.includeTOCCb.GetValue()
-        self.cfg.pdfShowTOC = self.showTOCCb.GetValue()
-        self.cfg.pdfOpenOnCurrentPage = self.openOnCurrentPageCb.GetValue()
-        self.cfg.pdfRemoveNotes = self.removeNotesCb.GetValue()
-        self.cfg.pdfOutlineNotes = self.outlineNotesCb.GetValue()
-        self.cfg.pdfShowMargins = self.marginsCb.GetValue()
-
-        self.outlineNotesCb.Enable(not self.cfg.pdfRemoveNotes)
-
-    def cfg2gui(self):
-        self.includeTOCCb.SetValue(self.cfg.pdfIncludeTOC)
-        self.showTOCCb.SetValue(self.cfg.pdfShowTOC)
-        self.openOnCurrentPageCb.SetValue(self.cfg.pdfOpenOnCurrentPage)
-        self.removeNotesCb.SetValue(self.cfg.pdfRemoveNotes)
-        self.outlineNotesCb.SetValue(self.cfg.pdfOutlineNotes)
-        self.marginsCb.SetValue(self.cfg.pdfShowMargins)
-
-
 class PDFFontsPanel(wx.Panel):
     def __init__(self, parent, id, cfg):
         wx.Panel.__init__(self, parent, id)
@@ -1650,6 +848,3 @@ class PDFFontsPanel(wx.Panel):
 
         return f.getPostscriptName()
 
-
-def cmpfunc(a, b):
-    return (a > b) - (a < b)
